+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.37.0] - 2022-09-55
+
+### Added
+
+- typed state support
+
+### Fixed
+
+- behavior of router configuration
+
 ## [3.36.1-2] - 2022-08-11
 
 ### Fixed